+++ ByteTrack/tutorials/centertrack/byte_tracker.py
@@ -15,7 +15,7 @@ class STrack(BaseTrack):
     def __init__(self, tlwh, score):
 
         # wait activate
-        self._tlwh = np.asarray(tlwh, dtype=np.float)
+        self._tlwh = np.asarray(tlwh, dtype=float)
         self.kalman_filter = None
         self.mean, self.covariance = None, None
         self.is_activated = False
@@ -184,7 +184,7 @@ class BYTETracker(object):
         detections = []
         detections_second = []
         
-        scores = np.array([item['score'] for item in results if item['class'] == 1], np.float32)
+        scores = np.array([item['score'] for item in results if item['class'] == 1], float)
         bboxes = np.vstack([item['bbox'] for item in results if item['class'] == 1])  # N x 4, x1y1x2y2
         
         remain_inds = scores >= self.args.track_thresh
@@ -354,7 +354,7 @@ def remove_fp_stracks(stracksa, n_frame=10):
     remain = []
     for t in stracksa:
         score_5 = t.score_list[-n_frame:]
-        score_5 = np.array(score_5, dtype=np.float32)
+        score_5 = np.array(score_5, dtype=float)
         index = score_5 < 0.45
         num = np.sum(index)
         if num < n_frame:
