+++ ByteTrack/tutorials/cstrack/byte_tracker.py
@@ -23,7 +23,7 @@ class STrack(BaseTrack):
     def __init__(self, tlwh, score):
 
         # wait activate
-        self._tlwh = np.asarray(tlwh, dtype=np.float)
+        self._tlwh = np.asarray(tlwh, dtype=float)
         self.kalman_filter = None
         self.mean, self.covariance = None, None
         self.is_activated = False
@@ -179,8 +179,8 @@ class BYTETracker(object):
         self.det_thresh = opt.conf_thres
         self.buffer_size = int(frame_rate / 30.0 * opt.track_buffer)
         self.max_time_lost = self.buffer_size
-        self.mean = np.array(opt.mean, dtype=np.float32).reshape(1, 1, 3)
-        self.std = np.array(opt.std, dtype=np.float32).reshape(1, 1, 3)
+        self.mean = np.array(opt.mean, dtype=float).reshape(1, 1, 3)
+        self.std = np.array(opt.std, dtype=float).reshape(1, 1, 3)
 
         self.kalman_filter = KalmanFilter()
         self.low_thres = 0.1
