+++ ByteTrack/tutorials/qdtrack/tracker_reid_motion.py
@@ -17,7 +17,7 @@ class STrack(BaseTrack):
     def __init__(self, tlwh, score, temp_feat, buffer_size=30):
 
         # wait activate
-        self._tlwh = np.asarray(tlwh, dtype=np.float)
+        self._tlwh = np.asarray(tlwh, dtype=float)
         self.kalman_filter = None
         self.mean, self.covariance = None, None
         self.is_activated = False
@@ -389,7 +389,7 @@ def remove_fp_stracks(stracksa, n_frame=10):
     remain = []
     for t in stracksa:
         score_5 = t.score_list[-n_frame:]
-        score_5 = np.array(score_5, dtype=np.float32)
+        score_5 = np.array(score_5, dtype=float)
         index = score_5 < 0.45
         num = np.sum(index)
         if num < n_frame:
